+++ api/models.py
@@ -1,0 +1,21 @@
+from pydantic import BaseModel, Field
+from typing import Optional
+
+class Task(BaseModel):
+    description: Optional[str] = Field(
+        'no description',
+        title='Task description',
+        max_length=1024,
+    )
+    completed: Optional[bool] = Field(
+        False,
+        title='Shows whether the task was completed',
+    )
+
+    class Config:
+        schema_extra = {
+            'example': {
+                'description': 'Buy baby diapers',
+                'completed': False,
+            }
+        }
