+++ api/routers/task.py
@@ -1,50 +1,16 @@
-# pylint: disable=missing-module-docstring, missing-function-docstring, missing-class-docstring biri biri no fuba, tola la, da ca
 import uuid
 
-from typing import Optional, Dict
+from typing import Dict
 
-from fastapi import FastAPI, HTTPException, Depends
-from pydantic import BaseModel, Field
+from fastapi import APIRouter, HTTPException, Depends
 
-from tasks import *
- 
-# pylint: disable=too-few-public-methods
-class Task(BaseModel):
-    description: Optional[str] = Field(
-        'no description',
-        title='Task description',
-        max_length=1024,
-    )
-    completed: Optional[bool] = Field(
-        False,
-        title='Shows whether the task was completed',
-    )
+from ..database import DBSession, get_db
+from ..models import Task
 
-    class Config:
-        schema_extra = {
-            'example': {
-                'description': 'Buy baby diapers',
-                'completed': False,
-            }
-        }
+router = APIRouter()
 
-
-tags_metadata = [
-    {
-        'name': 'task',
-        'description': 'Operations related to tasks.',
-    },
-]
-
-app = FastAPI(
-    title='Task list',
-    description='Task-list project for the **Megadados** course',
-    openapi_tags=tags_metadata,
-)
-
-@app.get(
-    '/task',
-    tags=['task'],
+@router.get(
+    '',
     summary='Reads task list',
     description='Reads the whole task list.',
     response_model=Dict[uuid.UUID, Task],
@@ -54,10 +20,8 @@ async def read_tasks(completed: bool = None, db: DBSession = Depends(get_db)):
     return db.read_tasks(completed)
         
   
-
-@app.post(
-    '/task',
-    tags=['task'],
+@router.post(
+    '',
     summary='Creates a new task',
     description='Creates a new task and returns its UUID.',
     response_model=uuid.UUID,
@@ -67,9 +31,8 @@ async def create_task(item: Task, db: DBSession = Depends(get_db)):
     return db.create_task(item)
 
 
-@app.get(
-    '/task/{uuid_}',
-    tags=['task'],
+@router.get(
+    '/{uuid_}',
     summary='Reads task',
     description='Reads task from UUID.',
     response_model=Task,
@@ -84,9 +47,8 @@ async def read_task(uuid_: uuid.UUID, db: DBSession = Depends(get_db)):
         ) from exception
 
 
-@app.put(
-    '/task/{uuid_}',
-    tags=['task'],
+@router.put(
+    '/{uuid_}',
     summary='Replaces a task',
     description='Replaces a task identified by its UUID.',
 )
@@ -100,9 +62,8 @@ async def replace_task(uuid_: uuid.UUID, item: Task, db: DBSession = Depends(get
         ) from exception
 
 
-@app.patch(
-    '/task/{uuid_}',
-    tags=['task'],
+@router.patch(
+    '/{uuid_}',
     summary='Alters task',
     description='Alters a task identified by its UUID',
 )
@@ -117,9 +78,8 @@ async def alter_task(uuid_: uuid.UUID, item: Task, db: DBSession = Depends(get_d
         ) from exception
 
 
-@app.delete(
-    '/task/{uuid_}',
-    tags=['task'],
+@router.delete(
+    '/{uuid_}',
     summary='Deletes task',
     description='Deletes a task identified by its UUID',
 )
